Added Authentication to Journal Endpoints

+++ src/main/java/com/example/api/service/JournalEntryService.java
@@ -3,6 +3,8 @@ package com.example.api.service;
 import java.util.List;
 import java.util.Optional;
 
+import javax.management.RuntimeErrorException;
+
 import org.bson.types.ObjectId;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
@@ -27,7 +29,7 @@ public class JournalEntryService {
             UserEntry user = userService.findByUserName(userName);
             JournalEntry saved = journalEntryRepository.save(journalEntry);
             user.getJournalEntries().add(saved);
-            userService.saveEntry(user);
+            userService.saveUser(user);
         } catch (Exception e) {
             throw new Exception(e);
         }
@@ -45,10 +47,23 @@ public class JournalEntryService {
         return journalEntryRepository.findById(Id);
     }
 
-    public void deleteByID(ObjectId Id, String userName) {
-        UserEntry user = userService.findByUserName(userName);
-        user.getJournalEntries().removeIf(x -> x.getId().equals(Id));
-        userService.saveEntry(user);
-        journalEntryRepository.deleteById(Id);
+    @Transactional
+    public boolean deleteByID(ObjectId Id, String userName) {
+        boolean removed = false;
+        try {
+            UserEntry user = userService.findByUserName(userName);
+            removed = user.getJournalEntries().removeIf(x -> x.getId().equals(Id));
+            if(removed) {
+                userService.saveUser(user);
+                journalEntryRepository.deleteById(Id);
+            }
+        } catch (Exception e) {
+            throw new RuntimeException("An Error occured while deleting the entry");
+        }
+        return removed;
     }
+
+    // public List<JournalEntry> findByUserName(String userName) {
+        
+    // }
 }
